requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
propcache==0.3.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.7.1
urllib3==2.4.0
yarl==1.19.0
